# api/openapi.yaml
openapi: 3.1.0
info:
  title: ame API
  description: Adaptive Mastery Engine — quiz & exam platform with agent identity.
  license:
    name: ''
  version: 0.1.0
paths:
...
  /v1/assessments/{id}:
    get:
      tags:
      - assessments
      operationId: get_assessment
      parameters:
      - name: id
        in: path
        description: Assessment ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Assessment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentDetail'
        '401':
          description: Missing or invalid token
        '404':
          description: Assessment not found
      security:
      - bearer_auth: []
    delete:
      tags:
      - assessments
      operationId: delete_assessment
      parameters:
      - name: id
        in: path
        description: Assessment ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Assessment deleted successfully
        '401':
          description: Missing or invalid token
        '404':
          description: Assessment not found
      security:
      - bearer_auth: []
    patch:
      tags:
      - assessments
      operationId: patch_assessment
      parameters:
      - name: id
        in: path
        description: Assessment ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssessmentRequest'
        required: true
      responses:
        '200':
          description: Assessment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        '401':
          description: Missing or invalid token
        '403':
          description: Not the owner
        '404':
          description: Assessment not found
        '422':
          description: Validation failed
      security:
      - bearer_auth: []
...
  /v1/plans:
    post:
      tags:
      - crate::http::plans
      operationId: create_plan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanBody'
        required: true
      responses:
        '201':
          description: Study plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyPlan'
        '401':
          description: Missing or invalid token
        '403':
          description: Token lacks required scope
        '422':
          description: Validation failed
      security:
      - bearer_auth: []
  /v1/plans/{id}:
    get:
      tags:
      - crate::http::plans
      operationId: get_plan
      parameters:
      - name: id
        in: path
        description: Plan id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Study plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyPlan'
        '401':
          description: Missing or invalid token
        '403':
          description: Token lacks required scope
        '404':
          description: Not found
      security:
      - bearer_auth: []
...
components:
  schemas:
...
    Assessment:
      type: object
      required:
      - id
      - title
      - mode
      - status
      - objectives
      - totalPoints
      - showResultsDuring
      - affectsRating
      - visibility
      - method
      - createdBy
      - createdAt
      - updatedAt
      properties:
        affectsRating:
          type: boolean
        compositionTrace: {}
        course:
          type:
          - string
          - 'null'
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        description:
          type:
          - string
          - 'null'
        durationMin:
          type:
          - integer
          - 'null'
          format: int32
        id:
          type: string
          format: uuid
        method:
          type: string
        mode:
          $ref: '#/components/schemas/AssessmentMode'
        objectives:
          type: array
          items:
            type: string
        passingPoints:
          type:
          - integer
          - 'null'
          format: int32
        showResultsDuring:
          type: boolean
        status:
          $ref: '#/components/schemas/AssessmentStatus'
        timeLimitSeconds:
          type:
          - integer
          - 'null'
          format: int32
        title:
          type: string
        totalPoints:
          type: integer
          format: int32
        updatedAt:
          type: string
          format: date-time
        visibility:
          $ref: '#/components/schemas/AssessmentVisibility'
    AssessmentDetail:
      allOf:
      - $ref: '#/components/schemas/Assessment'
      - type: object
        required:
        - sections
        - questions
        properties:
          questions:
            type: array
            items:
              $ref: '#/components/schemas/AssessmentQuestion'
          sections:
            type: array
            items:
              $ref: '#/components/schemas/AssessmentSectionDetail'
    AssessmentMode:
      type: string
      enum:
      - practice
      - graded
    AssessmentQuestion:
      type: object
      required:
      - id
      - kind
      - prompt
      - payload
      - points
      - status
      - orderIndex
      properties:
        codeSnippet: {}
        explanation:
          type:
          - string
          - 'null'
        id:
          type: string
          format: uuid
        kind:
          type: string
        orderIndex:
          type: integer
          format: int32
        payload: {}
        points:
          type: integer
          format: int32
        prompt:
          type: string
        status:
          type: string
    AssessmentSectionDetail:
      type: object
      required:
      - id
      - title
      - orderIndex
      - weight
      - itemsCount
      - questions
      properties:
        id:
          type: string
          format: uuid
        itemsCount:
          type: integer
          format: int32
        mix: {}
        orderIndex:
          type: integer
          format: int32
        questions:
          type: array
          items:
            $ref: '#/components/schemas/AssessmentQuestion'
        title:
          type: string
        weight:
          type: number
          format: double
    AssessmentStatus:
      type: string
      enum:
      - draft
      - active
      - archived
...
    AssessmentVisibility:
      type: string
      enum:
      - public
      - private
...
    CreatePlanBody:
      type: object
      required:
      - goal
      properties:
        goal:
          type: string
        lookbackDays:
          type:
          - integer
          - 'null'
          format: int64
...
    StudyPlan:
      type: object
      required:
      - id
      - user_id
      - goal
      - lookback_days
      - generated_at
      - weeks
      properties:
        generated_at:
          type: string
          format: date-time
        goal:
          type: string
        id:
          type: string
          format: uuid
        lookback_days:
          type: integer
          format: int64
        user_id:
          type: string
          format: uuid
        weeks:
          type: array
          items:
            $ref: '#/components/schemas/StudyPlanWeek'
    StudyPlanItem:
      type: object
      required:
      - kind
      - ref_id
      - hours_est
      properties:
        hours_est:
          type: number
          format: double
        kind:
          type: string
        ref_id:
          type: string
    StudyPlanWeek:
      type: object
      required:
      - week_num
      - focus
      - items
      properties:
        focus:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/StudyPlanItem'
        week_num:
          type: integer
          minimum: 0
...
    UpdateAssessmentRequest:
      type: object
      properties:
        description:
          type:
          - string
          - 'null'
        objectives:
          type:
          - array
          - 'null'
          items:
            type: string
        status:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/AssessmentStatus'
        title:
          type:
          - string
          - 'null'
        visibility:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/AssessmentVisibility'
...